### 2018-09-04 #926 @whedon
[PRE REVIEW]: Gravity: Estimation Methods for Gravity Models in R
**Submitting author:** @pachamaltese (<a href="http://orcid.org/0000-0003-1017-7574">Mauricio Vargas</a>)
**Repository:** <a href="https://github.com/pachamaltese/gravity" target ="_blank">https://github.com/pachamaltese/gravity</a>
**Version:** v0.8.2
