### 2018-09-11 #926 @pachamaltese
this is not physics!! it's economics

El mar., 11 de sep. de 2018 a la(s) 16:21, Yo Yehudi (
notifications@github.com) escribió:

> Still hunting for reviewers - the majority of reviewers in the JOSS
> reviewer list
> <https://docs.google.com/spreadsheets/d/1PAPRJ63yq9aPC1COLjaQp8mHmEq3rZUzwUYxTulyu78/edit#gid=856801822>
> tend to focus on bioinformatics rather than physics related subjects.
>
> @cole-brokamp <https://github.com/cole-brokamp>, would you be interested
> in reviewing this?
>
