### 2018-09-11 #926 @cole-brokamp
> @cole-brokamp, would you be interested in reviewing this?

Looks interesting, but I think the domain is outside of my area of expertise.  Sorry I couldn't be of more help!
