### 2018-09-24 #926 @yochannah
@pachamaltese I haven't had any responses from the reviewers I've emailed so far unfortunately, but I've emailed a few more so hopefully we'll have someone soon. In the meantime I noticed that a couple of the links in your reference section have what I'm guessing is a copy/paste error - showing https://doi.org twice, like this:
