### 2018-09-25 #926 @yochannah
@pachamaltese That looks mostly better! The first link is fixed, but the final link in the references section is still broken - it looks ok visually but returns an error when you click.
