### 2018-10-02 #926 @DiegoKoz
Hi @pachamaltese, its a pleasure!
As I told @yochannah, it's the first time I do a review, and especially of this kind, I'll try to do my best
