### 2018-10-09 #926 @DiegoKoz
Hi @pachamaltese, while looking for a second reviewer, let me make some comments about documentation.

It would be nice that the vignette does more than creating the dataset. It's great that you include the dataset from Head and Mayer, It would be fantastic if you make a vignette where you use it for the different functions, and which are the advantages from each way of estimating the gravity model. I don't think this is an indispensable criteria for accepting the paper, because the documentation inside the functions seems enough, but it would be a nice thing to have. At least a minimal example in the readme of the repo.

Another thing you should add to the Readme is the Community guidelines: There should be clear guidelines for third-parties wishing to contribute, report issues and seek support
