### 2018-10-09 #926 @pachamaltese
thanks a lot @DiegoKoz
of course I'm expanding the vignettes
