### 2018-10-09 #926 @DiegoKoz
Also, I don't understand why you choose to return de summary of the models instead of the object? From what I see, this is general for all the functions. Returning the object you fit would allow the users to predict, make intervals, etc. And also use the `moldr` and `broom` packages.
