### 2018-10-09 #926 @pachamaltese
@DiegoKoz I'm gonna add you to contributors!! The summary was to do exactly what the Stata package does, and you are right that "free translation" would be better
