### 2018-10-09 #926 @pachamaltese
The summary was to do exactly what the Stata package does, and you are right that "free translation" would be better
