### 2018-10-09 #926 @DiegoKoz
@pachamaltese I think adding the reviewer as a contributor would emerge a conflict of interests (right @yochannah ?) but thanks anyway!!

### minor detail
The example in `?nbpml()` is not done _the tidy way_ as the rest of the documentation.

### detail
Also with `nbpml()`, my computer just crashed (twice) because of the amount of computation needed. This is because of `MASS::glm.nb()` is computationally demanding, right? I'm using the examples in the documentation.
