### 2018-10-09 #926 @pachamaltese
@DiegoKoz ok, no worries!
yeah, glm eats a lot of ram, you can check inside `nbpml.R` and see the short example for CRAN (i.e. a test with a smaller dataset that takes <10 seconds)
