### 2018-10-09 #926 @DiegoKoz
I think it's more a matter with CPU rather than RAM.

The example inside nbpml.R runs, but gives lot's of warnings like  "non-integer x = 16.451930" which lend me to this: [stackoverflow thread](https://stats.stackexchange.com/questions/164889/how-to-deal-with-non-integer-warning-from-negative-binomial-glm )
the negative binomial is used for counting. That may be one of the reasons for the inefficiency with the computation. I assume that you didn't come up with using this distribution but you are rather reproducing it from the literature (I'm not sure from where).
If that's the case, I would suggest a warning in the documentation and removing the warnings from the output (because it is ok to give something that isn't an integer)
