### 2018-10-09 #926 @DiegoKoz
### Minor detail.

Take this into account:
![image](https://user-images.githubusercontent.com/28060531/46699645-d9b66900-cbf0-11e8-9072-4a7b24cb546a.png)

I change "USA" for a wrong name. The error should say something about the wrong name. Probably its filtering and ending with an empty tibble that afterward gives this error message. I should be easy to check that the names are ok, right?
