### 2018-10-10 #926 @paulofelipe
Nice to know about this package. I have one suggestion. Nowadays, one of the recommendations to assess the effects of trade policy is to use pair fixed effects ([wto book](https://www.wto.org/english/res_e/booksp_e/advancedwtounctad2016_e.pdf)). However, in some settings, this can mean the user will need a method that works with high dimensional fixed effects (in ols or ppml, for example). Otherwise, the estimation can be memory intensive and slow. After a quick look at the repository, I didn't notice any mention to that. In Stata, [Thomas Zylkin](http://www.tomzylkin.com/) wrote a specific function for the ppml case ([ppml_panel_sg](https://econpapers.repec.org/software/bocbocode/S458249.htm)). For ols estimation, it is common to use the `reg2hdfe` function.

In R, the user can estimate a high dimensional linear model using the [lfe package](https://cran.r-project.org/web/packages/lfe/lfe.pdf). In order to allow high dimensional fixed effects in PPML, I think it's necessary to rewrite ppml model using `lfe` inside. I have a [sketch here](https://github.com/paulofelipe/Structural-Gravity/blob/master/R/aux_functions.R).

Again, great to know about the package. I will try it soon.
