### 2018-10-13 #926 @pachamaltese
@DiegoKoz once again thanks a lot for the feedback !!
I did some changes to the functions so that you can now use broom with it, and here I added a vignette (WIP) http://pacha.hk/gravity/articles/crash-course-on-gravity-models.html that explains the models

I'll add the rest of the feedback such as warnings and more, and also to shorten the paper draft a bit.
