### 2018-10-21 #926 @yochannah
@DiegoKoz @pachamaltese You two are doing such a great job reviewing the package I'll kick off the formal review now. I'd still like a second reviewer but may add them in later in the process.
