### 2018-10-22 pachamaltese/gravity#4 @pachadotdev
nbpml
Hi @jbkunst

@DiegoKoz gladly provided us this feedback:

The example inside nbpml.R runs, but gives lot's of warnings like "non-integer x = 16.451930" which lend me to this: https://stats.stackexchange.com/questions/164889/how-to-deal-with-non-integer-warning-from-negative-binomial-glm

Provided you are the statistician here: What would be the best way to patch this?
