### 2018-10-22 pachamaltese/gravity#5 @pachadotdev
ppml
@paulofelipe also provided this valuable piece:

Nice to know about this package. I have one suggestion. Nowadays, one of the recommendations to assess the effects of trade policy is to use pair fixed effects (wto book). However, in some settings, this can mean the user will need a method that works with high dimensional fixed effects (in ols or ppml, for example). Otherwise, the estimation can be memory intensive and slow. After a quick look at the repository, I didn't notice any mention to that. In Stata, Thomas Zylkin wrote a specific function for the ppml case (ppml_panel_sg). For ols estimation, it is common to use the reg2hdfe function.

In R, the user can estimate a high dimensional linear model using the lfe package. In order to allow high dimensional fixed effects in PPML, I think it's necessary to rewrite ppml model using lfe inside. I have a sketch [here](https://github.com/paulofelipe/Structural-Gravity/blob/master/R/aux_functions.R).
