### 2018-10-22 pachamaltese/gravity#6 @DiegoKoz
documentation typo
Hi! I'm copying some previous comments.

in the documentation of the functions, the `regressors` param, it says "curreny", I think it is repeated in all the functions.

_Originally posted by @DiegoKoz in https://github.com/openjournals/joss-reviews/issues/926#issuecomment-428302360_
