### 2018-10-22 pachamaltese/gravity#7 @DiegoKoz
params order
This, in particular, would be an essential change and if you want to change it afterward it may imply some compatibility issues with older versions, so I really suggest to make this changes as soon as possible

---------

 It's kind of problematic that the function renames the first input from the `regressors` param, and always assume that is a distance. It would be better to add the parameter `distance`, as you added the `incomes` as a separate thing.  For example, I run the following:
`bvu(dependent_variable = "flow", regressors = c("flow","contig","rta"),
incomes = c("gdp_o", "gdp_d"), codes = c("iso_o", "iso_d"),
robust = TRUE, data = gravity_no_zeros)`

And it didn't make any warning.

In general, I'm not comfortable with the strict dependence of the ordering (inc_0, inc_d). I think it would be better to have different parameters for those things, to avoid confusion.

The function works as it is, and the documentation clarifies the needed order, so just take it as a suggestion.

_Originally posted by @DiegoKoz in https://github.com/openjournals/joss-reviews/issues/926#issuecomment-428302360_
