### 2018-10-22 pachamaltese/gravity#7 @pachadotdev
thanks again @DiegoKoz
it was fixed [here](https://github.com/pachamaltese/gravity/tree/1fcbd518b4c4e5d06f5544a847aa794d3c3ded52)

as well as BVW and DDM
the rest of the functions were updated in later commits
