### 2018-10-22 pachamaltese/gravity#8 @DiegoKoz
error messages
Similiar to [this](https://github.com/openjournals/joss-reviews/issues/926#issuecomment-428359179):

consider this
![image](https://user-images.githubusercontent.com/28060531/47310879-b33ff700-d60e-11e8-83ee-2f871de34653.png)

From the documentation example, I change the `filter_destination` for some wrong name. The error message is not very clear, as there is no problem with 'distw'.  I think this just happens with `filter_origin` and `filter_destination`, and could be easily  solved with some internal check of this names inside the functions.

In this case, for example, I would like to have an error message like " 'CHH' is not a valid countryname"
