### 2018-10-22 #1038 @DiegoKoz
> Hi! I'll open new issues in the [original repo](https://github.com/pachamaltese/gravity) with the comments I did in the previous repo, and link them here

https://github.com/pachamaltese/gravity/issues/8#issue-372641952
