### 2018-10-22 pachamaltese/gravity#4 @DiegoKoz
Hi @jbkunst !

[this](https://github.com/openjournals/joss-reviews/issues/926#issuecomment-428343346) was the original message.
