### 2018-10-22 pachamaltese/gravity#8 @pachadotdev
@DiegoKoz: thanks

@jbkunst: look, JOSS reviewers gave some cool feedback that will be added to this repo from now
do you think we can include your messageline() tricks around informative error messages?
beware github is acting funny so let's wait a bit until gh status informs otherwise
