### 2018-10-22 #1038 @DiegoKoz
> Hi! I'll open new issues in the [original repo](https://github.com/pachamaltese/gravity) with the comments I did in the previous repo, and link them here

[https://github.com/pachamaltese/gravity/issues/8](https://github.com/pachamaltese/gravity/issues/8)
[https://github.com/pachamaltese/gravity/issues/4](https://github.com/pachamaltese/gravity/issues/4)

--------
This are the remaining two. the rest of the suggestions are already implemented
