### 2018-10-22 #1038 @DiegoKoz
@pachamaltese please add a plain-text  license
