### 2018-10-22 #1038 @DiegoKoz
@pachamaltese please add __Community guidelines__:

> clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
