### 2018-10-22 #1038 @DiegoKoz
> @pachamaltese please add **Community guidelines**:
>
> > clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

You add them in README.rmd instead of README.md and it is not display in the github page
