### 2018-10-22 #1038 @DiegoKoz
> @yochannah just to inform that github became funnier again
> I added some feedback from @DiegoKoz and @paulofelipe to the repo, and therefore I updated the version to 0.9 after many iterations

I think this is the last thing that remains pending. @yochannah how do we proceed with this?
