### 2018-11-04 #1038 @yochannah
@DiegoKoz @pachamaltese I note there are a couple of issues in the gravity repo suggested by @DiegoKoz - do you feel like these need to be addressed before we accept?
