### 2018-11-05 #1038 @pachamaltese
Hi. Yes, please!
Let me tweak one of the functions a bit, if I am sleepless maybe I optimize
it a lot.
