### 2018-11-05 #1038 @pachamaltese
@DiegoKoz @yochannah I can always go back to 0.8.2
I just followed http://r-pkgs.had.co.nz/release.html#release-version
