### 2018-11-05 #1038 @DiegoKoz
>
>
> @DiegoKoz @yochannah I can always go back to 0.8.2
> I just followed http://r-pkgs.had.co.nz/release.html#release-version

yes, it's fine, but I don't understand if @yochannah should update de info of the review, or I just accept that item.

Besides de issue you are working on, the rest is fine for me and I think we could accept the review
