### 2018-11-12 #1038 @pachamaltese
@yochannah @DiegoKoz
Hi!! First of all, thanks a lot for the valuable feedback.
I closed all issues except https://github.com/pachamaltese/gravity/issues/5 after doing various improvements. PPML would requiere extensive coding and that issue is more performance related than user input's checks or statistical concerns.
After all the improvements I updated to v0.9.0 and, IMO, I consider the package to be ready.
