### 2018-11-18 #1038 @yochannah
@pachamaltese I've made a PR with a few minor grammar edits for the paper: https://github.com/pachamaltese/gravity/pull/12 - Oh wait, I see you have already merged it, nice!
