### 2018-11-18 #1038 @yochannah
@pachamaltese - At this point could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? Also take note of the authors list - it may pick my typo contribution up - you'll need to manually edit the metadata to remove me.
