### 2018-11-19 #1038 @arfon
> @yochannah my apologies for the delay, here is the upload with the source and binary build https://zenodo.org/record/1490783#.W_GhIuhKhPY

:wave: thanks for making the archive DOI. Unfortunately what you have done won't work. The DOI you have used with Zenodo is the DOI that _JOSS_ will issue (https://doi.org/10.21105/joss.01038). Instead, you need another DOI for the software archive.

When you have a new DOI, please notify us here and we'll proceed with accepting the paper.
