### 2019-02-12 c1f605d
creates function to drop observations as Romain suggested in #13
