+++ R/bvw.R
@@ -20,16 +20,16 @@
 #' the estimation of a gravity equation by \code{bvw} using panel data,
 #' we do not recommend to apply this method in this case.
 #'
-#' @param dependent_variable (Type: character) name of the dependent variable. This dependent variable is 
-#' divided by the product of unilateral incomes such (i.e. \code{income_origin} and \code{income_destination}) 
+#' @param dependent_variable (Type: character) name of the dependent variable. This dependent variable is
+#' divided by the product of unilateral incomes such (i.e. \code{income_origin} and \code{income_destination})
 #' and logged afterwards.
 #'
-#' @param distance (Type: character) name of the distance variable that should be taken as the key independent variable 
+#' @param distance (Type: character) name of the distance variable that should be taken as the key independent variable
 #' in the estimation. The distance is logged automatically when the function is executed.
 #'
 #' @param additional_regressors (Type: character) names of the additional regressors to include in the model (e.g. a dummy
-#' variable to indicate contiguity). Unilateral metric variables such as GDP should be inserted via the arguments 
-#' \code{income_origin} and \code{income_destination}. As country specific effects are subdued due to demeaning, no further 
+#' variable to indicate contiguity). Unilateral metric variables such as GDP should be inserted via the arguments
+#' \code{income_origin} and \code{income_destination}. As country specific effects are subdued due to demeaning, no further
 #' unilateral variables apart from incomes can be added.
 #'
 #' Write this argument as \code{c(contiguity, common currency, ...)}. By default this is set to \code{NULL}.
@@ -38,18 +38,18 @@
 #'
 #' @param income_destination (Type: character) destination income variable (e.g. GDP) in the dataset.
 #'
-#' @param code_origin (Type: character) country of origin variable (e.g. ISO-3 country codes). The variables are grouped 
+#' @param code_origin (Type: character) country of origin variable (e.g. ISO-3 country codes). The variables are grouped
 #' using this parameter.
 #'
-#' @param code_destination (Type: character) country of destination variable (e.g. country ISO-3 codes). The variables are 
+#' @param code_destination (Type: character) country of destination variable (e.g. country ISO-3 codes). The variables are
 #' grouped using this parameter.
-#' 
+#'
 #' @param robust (Type: logical) whether robust fitting should be used. By default this is set to \code{FALSE}.
 #'
 #' @param data (Type: data.frame) the dataset to be used.
 #'
 #' @param ... Additional arguments to be passed to the function.
-#' 
+#'
 #' @references
 #' For more information on gravity models, theoretical foundations and
 #' estimation methods in general see
@@ -63,7 +63,7 @@
 #' \insertRef{Baier2009}{gravity}
 #'
 #' \insertRef{Baier2010}{gravity}
-#' 
+#'
 #' \insertRef{Feenstra2002}{gravity}
 #'
 #' \insertRef{Head2010}{gravity}
@@ -105,7 +105,6 @@
 #'   robust = FALSE,
 #'   data = grav_small
 #' )
-#'
 #' @return
 #' The function returns the summary of the estimated gravity model as an
 #' \code{\link[stats]{lm}}-object.
@@ -143,24 +142,18 @@ bvw <- function(dependent_variable,
   stopifnot(is.character(code_origin), code_origin %in% names(data), length(code_origin) == 1)
   stopifnot(is.character(code_destination), code_destination %in% names(data), length(code_destination) == 1)
 
-  # Discarding unusable observations ----------------------------------------
-  d <- data %>%
[…]
