### 2019-02-12 4e92e83
adds Lionel feedback about git add .sym() simplification on bvu.R and discard_unusable.R
