+++ R/hm_summary.R
@@ -231,7 +231,7 @@ hm_summary <- function(model, robust = FALSE, ...) {
   }
 
   # Remove R-sq and F statistic for particular classes ----------------------
-  particular_class <- any(class(model) %in% paste0("gravity_", c("gpml", "nbpml", "nls")))
+  particular_class <- any(class(model) %in% paste0("gravity_", c("gpml", "nbpml", "ppml", "nls")))
 
   if (particular_class == TRUE) {
     ans$r.squared <- NULL
