### 2019-04-04 a3b041e
added JOSS badge and fixed hexicon
