+++ README.Rmd
@@ -7,12 +7,13 @@ bibliography: inst/REFERENCES.bib
 knitr::opts_chunk$set(echo = TRUE)
 ```
 
-# gravity <img src="https://pacha.hk/gravity/apple.svg" width=150 align="right" alt="sticker"/>
+# gravity <img src="https://pachamaltese.github.io/gravity/hexicon.svg" width=150 align="right" alt="sticker"/>
 
 [![Travis-CI Build Status](https://travis-ci.org/pachamaltese/gravity.svg?branch=master)](https://travis-ci.org/pachamaltese/gravity)
 [![cran checks](https://cranchecks.info/badges/summary/gravity)](https://cran.r-project.org/web/checks/check_results_gravity.html)
 [![Coverage status](https://codecov.io/gh/pachamaltese/gravity/branch/master/graph/badge.svg)](https://codecov.io/github/pachamaltese/gravity?branch=master)
-[![DOI](https://zenodo.org/badge/doi/10.17713/ajs.v47i4.688.svg)](https://doi.org/10.17713/ajs.v47i4.688)
+[![JOSS](http://joss.theoj.org/papers/10.21105/joss.01038/status.svg)](https://doi.org/10.21105/joss.01038)
+[![Austrian Journal of Statistics](https://zenodo.org/badge/doi/10.17713/ajs.v47i4.688.svg)](https://doi.org/10.17713/ajs.v47i4.688)
 
 ## About
 
