+++ README.md
@@ -1,63 +1,134 @@
 
-gravity <img src="https://pacha.hk/gravity/apple.svg" width=150 align="right" alt="sticker"/>
-=============================================================================================
+# gravity <img src="https://pachamaltese.github.io/gravity/hexicon.svg" width=150 align="right" alt="sticker"/>
 
-[![Travis-CI Build Status](https://travis-ci.org/pachamaltese/gravity.svg?branch=master)](https://travis-ci.org/pachamaltese/gravity) [![cran checks](https://cranchecks.info/badges/summary/gravity)](https://cran.r-project.org/web/checks/check_results_gravity.html) [![Coverage status](https://codecov.io/gh/pachamaltese/gravity/branch/master/graph/badge.svg)](https://codecov.io/github/pachamaltese/gravity?branch=master) [![DOI](https://zenodo.org/badge/doi/10.17713/ajs.v47i4.688.svg)](https://doi.org/10.17713/ajs.v47i4.688)
+[![Travis-CI Build
+Status](https://travis-ci.org/pachamaltese/gravity.svg?branch=master)](https://travis-ci.org/pachamaltese/gravity)
+[![cran
+checks](https://cranchecks.info/badges/summary/gravity)](https://cran.r-project.org/web/checks/check_results_gravity.html)
+[![Coverage
+status](https://codecov.io/gh/pachamaltese/gravity/branch/master/graph/badge.svg)](https://codecov.io/github/pachamaltese/gravity?branch=master)
+[![JOSS](http://joss.theoj.org/papers/10.21105/joss.01038/status.svg)](https://doi.org/10.21105/joss.01038)
+[![Austrian Journal of
+Statistics](https://zenodo.org/badge/doi/10.17713/ajs.v47i4.688.svg)](https://doi.org/10.17713/ajs.v47i4.688)
 
-About
------
+## About
 
-Gravity models are used to explain bilateral flows related to the sizes of bilateral partners, a measure of distance between them and other influences on interaction costs.
+Gravity models are used to explain bilateral flows related to the sizes
+of bilateral partners, a measure of distance between them and other
+influences on interaction costs.
 
-The underlying idea is rather simple. The greater the masses of two bodies and the smaller the distance between them, the stronger their attraction. This concept is applied to several research topics such as trade, migration or foreign direct investment. Even though their basic idea is rather simple, gravity models can become complex regarding the choice of models or estimation methods.
+The underlying idea is rather simple. The greater the masses of two
+bodies and the smaller the distance between them, the stronger their
+attraction. This concept is applied to several research topics such as
+trade, migration or foreign direct investment. Even though their basic
+idea is rather simple, gravity models can become complex regarding the
+choice of models or estimation methods.
 
-As especially for gravity beginners it is difficult to get an overview of the different methods and implement them in R, the package provides a wrapper of different standard estimation methods for gravity models. By considering the descriptions and codes of the estimation methods, users get a comprehensive and application-oriented overview of the different methods, see which method may be suitable for a certain research question or type of data and extend the code available to their research projects.
+As especially for gravity beginners it is difficult to get an overview
+of the different methods and implement them in R, the package provides a
+wrapper of different standard estimation methods for gravity models. By
+considering the descriptions and codes of the estimation methods, users
+get a comprehensive and application-oriented overview of the different
+methods, see which method may be suitable for a certain research
+question or type of data and extend the code available to their research
+projects.
 
-The package provides estimation methods for log-log models and multiplicative models.
+The package provides estimation methods for log-log models and
+multiplicative models.
 
 Log-log estimation covers:
 
--   Ordinary Least Squares (OLS)
[…]
