+++ DESCRIPTION
@@ -3,12 +3,27 @@ Type: Package
 Title: Estimation Methods for Gravity Models
 Version: 0.9.7
 Authors@R: c(
-  person("Anna-Lena", "Woelwer", role = c("aut"), email = "woelwer@uni-trier.de"),
-  person("Jan Pablo", "Burgard", role = c("aut"), email = "burgardj@uni-trier.de"),
-  person("Joshua", "Kunst", role = c("aut"), email = "jbkunst@gmail.com"),
-  person("Mauricio", "Vargas", role = c("aut", "cre"), email = "mvargas@dcc.uchile.cl"),
-  person("Romain", "Francois", role = c("ctb"), email = "romain@purrple.cat"),
-  person("Lionel", "Henry", role = c("ctb"), email = "lionel@rstudio.com")
+  person("Anna-Lena", "Woelwer",
+    role = "aut",
+    email = "woelwer@uni-trier.de"),
+  person("Jan Pablo", "Burgard",
+    role = "aut",
+    email = "burgardj@uni-trier.de"),
+  person("Joshua", "Kunst", 
+    role = "aut",
+    email = "jbkunst@gmail.com"),
+  person("Mauricio", "Vargas", 
+    role = c("aut","cre","cph"), 
+    email = "mvargas@dcc.uchile.cl", 
+    comment = c(ORCID = "0000-0003-1017-7574")),
+  person("Romain", "Francois", 
+    role = "ctb", 
+    email = "romain@purrple.cat",
+    comment = "adapted parts of the code to use dplyr 0.8.0"),
+  person("Lionel", "Henry", 
+    role = c("ctb"), 
+    email = "lionel@rstudio.com",
+    comment = "simplified parts of the code")
   )
 Description: A wrapper of different standard estimation methods for gravity models. 
   This package provides estimation methods for log-log models and multiplicative models.
