### 2019-04-08 pachamaltese/gravity#5 @zauster
The [alpaca package](https://github.com/amrei-stammann/alpaca) can fit highly dimensional fixed effects in glm's.
I have used it several times, it is really blazing fast.
